Bug 1826568 - Move audioPlaybackSystemNotificationTest to MediaNotificationTest class
